+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Created an `Xlsxir.TableId` module which controls an agent process that temporarily holds a table identifier during the extraction process.
 - Refactored `Xlsxir` access functions to work with `Xlsxir.multi_extract/3` whereby a table identifier is passed through the various functions to specify which ETS process is to be accessed. 
 - Refactored `Xlsxir.SaxParser`, `Xlsxir.ParseWorksheet` and `Xlsxir.Worksheet` modules to support new functionality.
+- Refactored `Xlsxir.ParseWorksheet` to ignore empty cells.
 - Updated documentation and tests
 - Fixed a few minor bugs that were generating warning messages.